+++ MigrationProject/user.c
@@ -35,7 +35,7 @@ void MRSG_set_dfs_f ( void (*f)(char** dfs_matrix, size_t chunks, size_t workers
     user.dfs_f = f;
 }
 
-void MRSG_set_map_output_f ( size_t (*f)(size_t mid, size_t rid) )
+void MRSG_set_map_output_f ( unsigned long long (*f)(size_t mid, size_t rid) )
 {
     user.map_output_f = f;
 }
